+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/DefaultFieldFactory.java
@@ -22,7 +22,7 @@ import is.codion.common.model.table.ColumnConditionModel;
 import is.codion.common.value.Value;
 import is.codion.common.value.ValueSet;
 import is.codion.swing.common.ui.component.builder.ComponentBuilder;
-import is.codion.swing.common.ui.component.table.ColumnConditionPanel.FieldFactory;
+import is.codion.swing.common.ui.component.table.FilterColumnConditionPanel.FieldFactory;
 
 import javax.swing.JComponent;
 import java.math.BigDecimal;